Stop frame descriptor overflow
A port of https://github.com/ocaml-flambda/ocaml/commit/9db53f1b.

+++ ocaml/runtime/caml/stack.h
@@ -87,9 +87,9 @@ struct caml_context {
 
 typedef struct {
   uintnat retaddr;
-  unsigned short frame_size;
-  unsigned short num_live;
-  unsigned short live_ofs[1 /* num_live */];
+  unsigned int frame_size;
+  unsigned int num_live;
+  unsigned int live_ofs[1 /* num_live */];
   /*
     If frame_size & 2, then allocation info follows:
   unsigned char num_allocs;
